+++ src/xulbux/xx_string.py
@@ -22,18 +22,9 @@ class String:
         """Replaces all special space characters with normal spaces.
         Also replaces tab characters with `tab_spaces` spaces."""
         return (
-            string.replace("\t", " " * tab_spaces)
-            .replace("\u2000", " ")
-            .replace("\u2001", " ")
-            .replace("\u2002", " ")
-            .replace("\u2003", " ")
-            .replace("\u2004", " ")
-            .replace("\u2005", " ")
-            .replace("\u2006", " ")
-            .replace("\u2007", " ")
-            .replace("\u2008", " ")
-            .replace("\u2009", " ")
-            .replace("\u200A", " ")
+            string.replace("\t", " " * tab_spaces).replace("\u2000", " ").replace("\u2001", " ").replace("\u2002",
+            " ").replace("\u2003", " ").replace("\u2004", " ").replace("\u2005", " ").replace("\u2006",
+            " ").replace("\u2007", " ").replace("\u2008", " ").replace("\u2009", " ").replace("\u200A", " ")
         )
 
     @staticmethod
@@ -44,13 +35,8 @@ class String:
         inside of. So if your string will be `"string"`, you should pass `"` to the parameter
         `str_quotes`. That way, if the string includes the same quotes, they will be escaped."""
         string = (
-            string.replace("\\", r"\\")
-            .replace("\n", r"\n")
-            .replace("\r", r"\r")
-            .replace("\t", r"\t")
-            .replace("\b", r"\b")
-            .replace("\f", r"\f")
-            .replace("\a", r"\a")
+            string.replace("\\", r"\\").replace("\n", r"\n").replace("\r", r"\r").replace("\t", r"\t").replace("\b",
+            r"\b").replace("\f", r"\f").replace("\a", r"\a")
         )
         if str_quotes == '"':
             string = string.replace(r"\\'", "'").replace(r'"', r"\"")
@@ -77,10 +63,8 @@ class String:
     @staticmethod
     def decompose(case_string: str, seps: str = "-_", lower_all: bool = True) -> list[str]:
         """Will decompose the string (any type of casing, also mixed) into parts."""
-        return [
-            (part.lower() if lower_all else part)
-            for part in _re.split(rf"(?<=[a-z])(?=[A-Z])|[{_re.escape(seps)}]", case_string)
-        ]
+        return [(part.lower() if lower_all else part)
+            for part in _re.split(rf"(?<=[a-z])(?=[A-Z])|[{_re.escape(seps)}]", case_string)]
 
     @staticmethod
     def to_camel_case(string: str, upper: bool = True) -> str:
@@ -118,4 +102,4 @@ class String:
     @staticmethod
     def split_count(string: str, count: int) -> list[str]:
         """Will split the string every `count` characters."""
-        return [string[i : i + count] for i in range(0, len(string), count)]
+        return [string[i:i + count] for i in range(0, len(string), count)]