+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   the syntax highlighting colors and styles can be customized via the new param <code>syntax_highlighting: dict[*str*, *str*] = {...}</code>
 * added two new methods `Data.serialize_bytes()` and `Data.deserialize_bytes()`
 * made the method `String.to_type()` be able to also interpret and convert large complex structures
+* added the new parameter <code>strip_spaces: *bool* = True</code> to the method `regex.brackets()` which makes it possible to not ignore spaces around the content inside the brackets
 
 ## 29.01.2025 `v1.6.5`
 * now the method `FormatCodes.to_ansi()` automatically converts the param `string` to a *`str`* if it isn't one already